requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
numpy>=1.24.0

